Extract runner module
