+++ lib/mix/tasks/learn.ex
@@ -6,7 +6,7 @@ defmodule Mix.Tasks.Learn do
   @shortdoc "Learn Elixir through meditation"
 
   def run(_) do
-    System.at_exit(fn 0 -> Koans.run end)
+    System.at_exit(fn 0 -> Koans.Runner.run end)
     Koans.Lessons.load
   end
 end
